requirements.txt
prompt_toolkit>=3.0.43
pydantic>=2.7.0
pyyaml>=6.0.1
cryptography>=42.0.0
